Added new class to read Tahoe results (geo/run).
Added function to describe which variables are used by an element group. This
shoule be used in conjunction with Read***Variables which returns an array of
variable values for all variable labels of that type found in the database. Also
used with Read***Labels which returns a list of all variable labels of that
type found in the database.